+++ src/Examples/Settings/README.md
@@ -1,0 +1,4 @@
+# Settings Examples
+
+- [Adding Settings](./AddingSettings.cs)
+- [Consuming Settings](./ConsumingSettings.cs)
